Refactored problems table
