+++ leetcode-tracker/app/_components/tables/problem-table/ProblemTableWithActions.tsx
@@ -1,0 +1,27 @@
+'use client'
+import { useState } from "react";
+import { ProblemType } from "@/app/_types/types";
+import ProblemTable from "./ProblemTable";
+import Toggle from "../../buttons/Toggle";
+
+function Foo( {problems} : {problems: ProblemType[]} ) {
+
+  const [displayCompleted, setDisplayCompleted] = useState(true);
+
+  return (
+    <div className="flex-grow">
+    <div className="flex flex-row">
+      <Toggle
+        text='Show Completed'
+        isChecked={displayCompleted}
+        toggleAction={()=>{setDisplayCompleted(!displayCompleted)}}/>
+    </div>
+    <div className="w">
+      <ProblemTable problems={problems} displayCompleted={displayCompleted}/>
+    </div>
+  </div>
+  );
+}
+
+export default Foo;
+
